refactored tts
