+++ src/app/core/api/interfaces/tts.interface.ts
